Add jlog_ctx_bulk_read_messages Function
Add a new function, jlog_ctx_bulk_read_messages, which will allow
passing in a count and will attempt to read in <count> messages.

This will fail if not all messages are read properly.

This assumes that enough memory has been passed in to accomodate all the
messages; behavior is undefined if insufficient memory has been
allocated.