Fix Potential Short Write On jlog_file_pwritev (#62)
jlog_file_pwritev was not checking to verify that the correct number of
bytes were written. Change this so that if we get short writes, we
will retry until we actually write the whole amount.

Added a new function, jlog_file_pwritev_verify_return_value that takes the
expected bytes as an input; this is because otherwise, the existing
jlog_file_writev will have to loop through all of the vectors and calculate
the value when it's possible that this value will be known in advance. This
prevents this extra calculation when the correct value is already known without
breaking compatibility with the existing API.